+++ src/test/java/hwr/oop/PlanetTest.java
@@ -1,2 +1,13 @@
-package hwr.oop;public class PlanetTest {
+package hwr.oop;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class PlanetTest {
+    @Test
+    void generateSamplePlanet() {
+        int planetSize = 10;
+        Planet mars = new Planet(planetSize);
+        Assertions.assertThat(mars.getArea()).isEqualTo(100);
+    }
 }
